+++ doc/documents/changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 **May 12th, 2020**
 - added API call to update name, request is private at `/api/v1/user/name`
+- documentation server is now served by main server
 
 **May 11th, 2020**
 - Added Error Message to Login callback